Fix indentation in slime-repl in some cases
It was broken when there were some parens in the output (such as from logging or
cl:format) in the repl buffer above the current prompt. Fix this by only parsing
code that is below the prompt - it sounds more correct anyway. It fixes two
cases: 1st - When indenting the first line on the same line as the prompt. 2nd -
When indenting code below the prompt line that would be incorrectly indented to
a deeper level because of some stray parens up above.